refactor: move ReportAttachments route into routes/report
